main: Avoid pack/unpack fast paths if base internal format != dst format
In general, if the base internal format and the selected dst format are
different we can't be sure that directly packing or unpacking to the destination
format will produce correct results. In these cases we probably want to fall
back to other paths (like mesa_swizzle_and_convert) that can handle these
situations better.

An example of this includes a luminance internal format for which the driver
decided to use something like BGRA. In these case, unpacking to BGRA won't
match the requirements of the luminance internal format.

In the future we may want to allow these fast paths for specific combinations
of base internal format and dst format that are known to work in this way.