main: Avoid pack/unpack fast paths if base internal format != base format
In general, if the dst base internal format and the selected dst format are
different we can't be sure that directly packing or unpacking to the destination
format will produce correct results. In these cases we probably want to fall
back to other paths (like mesa_swizzle_and_convert) that can handle these
situations better.

An example of this includes a luminance internal format for which the driver
decided to use something like BGRA. In these case, unpacking to BGRA won't
match the requirements of the luminance internal format.

In the future we may want to allow these fast paths for specific cases
where we know the direct pack/unpack functions will do the right thing.