main: Add _mesa_swap2_copy and _mesa_swp4_copy
We have _mesa_swap{2,4}, but these do in-place byte-swapping only. The new
functions receive an extra parameter so we can swap bytes on a source
input array and store the results in a (possibly different) destination
array.

This is useful to implement byte-swapping in pixel uploads, since in this
case we need to swap bytes on the src data which is owned by the
application and we can't do an in-place swap.