glsl: fail when a shader's input var has not an equivalent out var in previous
GLSL ES 3.00 spec, 4.3.10 (Linking of Vertex Outputs and Fragment Inputs),
page 45 shows a table that explain that: when a Fragment Shader input is
declared and static used but there is no reference of inside vertex shader, a
link error should be returned.

Similar table or content is present in other GLSL specs: GLSL 4.40, GLSL 1.50,
etc but for more stages (vertex and geometry shaders, etc).

This patch detects that case and returns a link error.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalia.com>