i965/fs: initialize gl_FragData or a user-defined FS output
Before doing any assignment, initialize the FS ouput values to zero.

This way we avoid the existing noise when the shader assigns values
to less number of components than the output variable's type has.

NOTE TO MYSELF: This might not be what we expect to occur.
Check with the documentation if we can initialize FS outputs with
zeros or any other value and that it doesn't break anything else.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalia.com>