mesa: fix deletion of inactive bound transform feedback object
When a transform feedback object is bound and not active, the OpenGL ES
3.0 and GL_ARB_transform_feedback2 specs don't explicitly disallow its
deletion. Only the deletion of the default framebuffer object is
forbidden.

This patch follows what it is done for glDeleteTextures(), i.e.
the binding reverts to 0 (the default framebuffer object).

Fixes:

dEQP-GLES3.functional.lifetime.delete_bound.transform_feedback

Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalia.com>