mesa: Add new IR node ir_ssbo_store
Shader storage buffer objects (SSBO) require special handling: when we
detect writes to any channel of a shader buffer variable we need to
emit the corresponding write to memory. We will later add a lowering pass
that detects these writes and injects ir_ssbo_store nodes in the IR so
drivers can  generate code for the memory writes.

Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>