glsl: Do constant folding on ir_ssbo_store
We have ir_rvalues for various fields within ir_ssbo_store, so we want
constant folding to know about them. Specifically, the offset of the
store operation can be computed as the addition of a constant offset
and a constant base offset, so we want that constant addition replaced
with its constant result.