nir-vec4: Fix handling of constant values
The generation of immediate registers directly in get_nir_src() was buggy.

It turns out that the offset() function takes into account the register's declared
size, and that makes sense. But since we are declaring the register for the
constants with the full size of the ssa variable, then the call to offset() was
expanding too long into the next registers.