glsl: process interface packing qualifier for struct definition
This patch also propagates interface packing qualifier to struct's members
and to array of structs.

The created glsl_types are meant to replace the already defined in
former struct global definition. This replacement takes into account
that the field names cannot be different as it is the same struct
definition but with different packing.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalia.com>