i965/fs: Import array utils for the surface message builder.
Define a few transformations on register arrays which will be used
frequently during the construction of typed and untyped surface
message payloads.  Their purpose is simple but the implementation is
rather messy, so it makes a lot of sense to factor them out as
separate functions.

v2: Drop VEC4 suport.