i965/nir/vec4: Add get_nir_dst() and get_nir_src() methods
These methods are essential for the implementation of the NIR->vec4 pass. They
work similar to their fs_nir counter-parts.

When processing instructions, these methods are invoked to resolve the
brw registers (source or destination) corresponding to the NIR sources
or destination. It builds a map of NIR register index to brw register for
all registers locally allocated in a block.

In the case of get_nir_src(), it also builds immediate registers on-the-fly
when queried for a SSA source which at this point can only correspond to
constant values.