i965/nir/vec4: Add nir_swizzle_result() method to vec4_visitor
This is the NIR->vec4 version of vec4_visitor::swizzle_result(). It is
essentially the same code except that we also pass the destination register
as argument, so that the result is MOVed to that register instead of
to vec4_visitor::result, which is what vec4_visitor::swizzle_result() does
and would cause an extra MOV in our case.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=89580