i965/nir/vec4: Add skeleton implementation of nir_emit_texture()
This sets up the basic structure and placeholders for the implementation of
texture operations, which will be added incrementally in subsequent patches.

Apart from helping split the texture support into smaller patches, this patch
is useful to get a simplified picture of the steps involved in the emission of
texture operations.

(@TODO: apinheiro to review it)

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=89580