i965/nir: dot not assign direct uniform locations first for vec4-based shaders
In the vec4 backend we want uniform locations to be assigned consecutively
since that way the offsets produced by nir_lower_io are exactly what we need
to implement nir_intrinsic_load_uniform. Otherwise we would need a mapping to
match the output of nir_lower_io to the actual uniform registers we need to
use.