i965/vec4/nir: duplicate swizzles and writemasks for doubles
In align16 mode, swizzles and writemasks are always specified in terms
of 32-bit chunks, which means that for doubles, a swizzle like "yx" in
GLSL and NIR needs to be expanded to a swizzle like "zwxy". There are
two possible places to do this:

1. In the visitor when converting to the vec4 IR.
2. In the generator.

We adopt the first approach, which simplifies per-channel liveness
analysis since channels are always 32 bits.

v2: Assert on invalid writemasks. We don't expect to see writemasks on
    64-bit results other than XY, so assert on that. Otherwise we end up
    generating invalid writemaks that are detected later at codegen time,
    when it is more difficult to debug. We might also want to do something
    similar with swizzles but that would require that callers always set
    correct swizzle masks for channels XY, even if only channel X or Y
    are being used. We don't do this at the moment (Iago)

v3: Fix build after rebase. Call to brw_swizzle_for_nir_swizzle in
    optimize_predicate() needs to pass the bit_size parameter (Samuel).

Signed-off-by: Iago Toral Quiroga <itoral@igalìa.com>
Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalìa.com>