nir: handle different bit sizes when constant folding
v2: Use the bit-size information from the opcode information if defined (Iago)

v3: Use helpers to get type size and base type of nir_alu_type enum (Sam)

v4: Do not fallback to sized types to guess bit-size information. (Jason)

Signed-off-by: Iago Toral Quiroga <itoral@igalia.com>

FIXME: This should be squashed into the previous commit so we don't break
the build. The break happens because the python script that generates the
constant folding pass does not know how to handle the sized types introduced
by the previous commit until this patch, so it ends up generating code with
invalid types. Keep it separated for review purposes.