i965/vec4: avoid dependency control around Align1 instructions
It appears that not only math instructions, but also MOV_BYTES and
d2f/f2d or any instruction that uses Align1 mode cannot be in the middle
of a dependency control sequence or the GPU will hang (at least on my
BDW). This fixes GPU hangs on some fp64 tests.