Revert "XXX remove exec size hack"
That broke the execution size for smaller instructions and was the cause
of multiple GPU hangs observed in a normal piglit run (without fp64 tests)

FIXME: this, however, breaks fp64, since we produce instructions with a
dst.width of 4 that need an execution size of 8 for a SIMD8 execution.
Unfortunately, we can't just check for the type of the dst and ignore this
if it is a DF, because some times we will generate UD instructions to
operate on the DF hi/lo 32-bit components.

Later patches will address this.