Fix register types for 32-bit
The commit "i965/fs: use the NIR bit size when creating registers"
changed the types used to create some 32-bit vgrfs under
the suppositon that it should not really matter, but it does. That
broke a number of tests that would hit assertions because hw_reg
types would not match the vgrf type. If we remove the assertion the
tests fail, so we really need to keep the original register types here
or find why they matter and fix that.

Fixes crashes:
spec/arb_gpu_shader5/arb_gpu_shader5-interpolate*