nir: Make vecN operations unsigned so they do not get source modifiers
This is a partial revert of:

"nir: update opcode definitions for different bit sizes"

That change was done so that we didn't produce 64-bit integer instructions,
which at one point the constant folding infrastructure couldn't support,
but now it can so this is no longer necessary. Also, this created a problem
because lower_vecs_to_movs (or any other pass that needs to break these into
individual MOVs) would then need to know if the source operand was
integer or float to emit imov or fmov instructions in each case, something
that wasn't happening at the moment.

This fixes > 160 regressions.