mesa/formatquery: Check the resource support for INTERNALFORMAT_{X}_{SIZE,TYPE} pnames
Although the ARB_internalformat_query2 spec does not say that the /resource/ support should
be checked for the INTERNALFORMAT_{X}_{SIZE, TYPE} pnames, it is also true that it says that
the expected value for those pnames is the one returned by glGetTexLevelParameter and for this
function, if a certain combination of texture target and internalformat is not supported,
glGetTexLevelParameter will return the 'unsupported' response (zero or GL_NONE, depending on
the query). So we check the resource support although the extension spec does not say it
explicitly.