i965/fs: add SHUFFLE_32BIT_LOAD_RESULT_TO_64BIT_DATA helper
There will be a few places where we need to shuffle the result of a 32-bit
load into valid 64-bit data, so extract this logic into a separate helper
that we can reuse.

The shuffling needs to operate with WE_all set because we are changing the
layout of the data across the various channels. Otherwise we will run into
problems in non-uniform control-flow scenarios.

v2: Use subscript() instead of stride() (Curro)