i965/fs: add SHUFFLE_32BIT_DATA_FOR_64BIT_WRITE helper
This does the inverse operation of SHUFFLE_32BIT_LOAD_RESULT_TO_64BIT_DATA
and we will use it when we need to write 64-bit data in the layout expected
by untyped write messages.

Again, this needs to operate with WE_all set for the same reasons as the
inverse operation.