nir/lower_vec_to_movs: don't produce too large double operations via colaescing
If we see a 32-bit vec4 operation we might try to coalesce it, however,
if the 32-bit sources of this operation are in fact the result of an alu
operation producing 32-bit results from 64-bit operands, we can end up
with a 64-bit alu operation with more than 2 double operands. For example:

vec2 32 ssa_4 = d2f ssa_2
vec4 32 ssa_5 = vec4 ssa_4, ssa_4.y, ssa_4, ssa_4.y

could be coalesced into:

vec4 32 r0 = d2f ssa_2.xyxy

Which produces a d2f instruction operating on 4 doubles. We don't want this
if the split_doubles option has been set, so before we coalesce, check for
the bit sizes of the inputs and outputs of the source alu operation and
avoid coalescing in this case.

v2 (Sam)
- Use helper to get type size from nir_alu_type enum.