i965/vec4: fix UBO loads for doubles up to dvec2
Consider that the loads operate in units of 32-bits so for a 64-bit load
we need to read twice the amount of components. This means that we can
load up to 2 dvec components with single message.

If we have anything larger than a dvec2 we will have to emit a separate
message for the second part of the vector. A later commit will address this.