i965/fs: fix copy/constant propagation regioning checks
We were not accounting for reg_suboffset in the check for the start
of the region. Also, fs_reg::regs_read() already takes the stride
into account, so we should not multiply its result by the stride
again.

Incidentally, the stride fix opens more possibilities for copy-propagation
which uncovered another bug where we can copy-propagate an entry that has
been partially invalidated (only fp64 code seems affected by this according
to piglit). The next patch will fix this.