i965/fs: fix copy/constant propagation regioning checks
We were not accounting for reg_suboffset in the check for the start
of the region. This meant that would allow copy-propagation even if
the dst wrote to sub_regoffset 4 and our source read from
sub_regoffset 0, which is not correct. This was observed in fp64 code,
since there we use reg_suboffset to select the high 32-bit of a double.

Also, fs_reg::regs_read() already takes the stride into account, so we
should not multiply its result by the stride again. This was making
copy-propagation fail to copy-propagate cases that would otherwise be
safe to copy-propagate. Again, this was observed in fp64 code, since
there we use stride > 1 often.

Incidentally, these fixes open up more possibilities for copy propagation
which uncovered new bugs in copy-propagation. The folowing patches address
each of these new issues.

v2 (Sam):
- Rename function and add comment (Jason).
- Add an assert (Jason).