i965/fs/lower_simd_width: fix result transposition
The transposition needs to set exec_all() but it writes directly to the
original instruction's destination, which can lead to execmasking
problems if the original instruction did not have force_writemask_all
set. In that case, write the result of the operation to a temporary,
then move the result to the original instruction's destination without
forcing exec_all().