glsl: Copy propagate in array index function parameters
Copy propagate is not applied in function parameters when they are out
or inout.

But if the parameter is an array, we can copy propagate the index array.

This also fixes shaders@out-parameter-indexing piglit tests, that
exposes a wrong handling of inout function parameters in Mesa.

This commit does not produce any piglit regression, nor any change in
shader-db results.

v2:
  - Deal with swizzles (jasuarez)
  - Use a loop top peel off all array dereferences (Ilia)

Reviewed-by: Lars Hamre <chemecse@gmail.com>