mesa/glformats: add desktop gl checks on _mesa_base_tex_format
There are several internalformat that are not supported on gl es, so
it should return -1 if that is the case. This is needed in order to
get ARB_internalformat_query2 implementation deciding correctly if
a resource is supported or not on opengl es.

FWIW, in some cases, _mesa_base_fbo_format has equivalent checks
for those internalformats, although for this method it is implemented
as a check/break in most cases, to keep consistency within the function.

Acked-by: Eduardo Lima <elima@igalia.com>
Acked-by: Antia Puentes <apuentes@igalia.com>