i965/vec4: allow the vec4 IR to express a non-default execution size
In the vec4 backend the generator sets the execution size for all
instructions to 8, however, we will have to split certain DF instructions
to have an execution size of 4 in some cases, so we need to indicate
this explicitly so the generator can set the right execution size
for them.

An exec_size of 0 indicates that the instruction should use the default
execution size.

FIXME: Maybe just make the default 8 and have the generator always call
set_default_exec_size(inst->exec_size);

FIXME: this will be useful when we implement the vec4 SIMD splitting pass.