i965/vec4: Use <4,4,1> region for the source of double-to-single
Because the vec4 backend sets a vstride of 2 on DF operands, we
need to either go with a <2,2,1> or <4,4,1> region, but not
<2,4,1>. We can use <4,4,1> here since this executes in Align1
mode.

FIXME: should we merge this in the original implementation? I guess
it will depend on when in the series we decide to start using vstride=2
for DF operands.