i965/vec4: force vstride0 in more cases
A source region like <2,2,1>.xyxy:DF selects XXZZ of a dvec4. If we
have code such as:

mov g2.z g4.x

that creates a problem, because we end up writing g4.z in g2.z. To fix
this we want to generate a XXXX region and we can do that by exploiting
again the vstride=0 behavior of the hardware.

Since we are using a vstride=0 also when we need to read components Z
and W, add a force_vstride0 field in src_reg that we can set when we want
to exploit this behavior and set it to true in both scenarios.

FIXME: we probably want to make vstride=0 all the time, at least while
we rely on full scalarization.