i965/vec4: add a SIMD lowering pass
When operating with double-precision instructions we have some cases
where we need to use an exec_size of 4 to avoid running into certain
hardeare limitations and bugs. For example, the limitation that when
we write more than 1 register we also need to read more than 1
register affects instructions that read from uniforms.